Add checkbox to embed custom navigation links below the main table.

+++ app.py
@@ -475,7 +475,8 @@ def add_nav_link():
         url=data['url'],
         icon_class=data.get('icon_class', 'fas fa-link'),
         user_id=current_user.id,
-        order=new_order
+        order=new_order,
+        embed=data.get('embed', False)
     )
     db.session.add(link)
     db.session.commit()
@@ -484,17 +485,39 @@ def add_nav_link():
         'id': link.id,
         'name': link.name,
         'url': link.url,
-        'icon_class': link.icon_class
+        'icon_class': link.icon_class,
+        'embed': link.embed
     })
 
-@app.route('/api/nav_links/<int:link_id>', methods=['DELETE'])
+@app.route('/api/nav_links/<int:link_id>', methods=['PUT', 'DELETE'])
 @login_required
-def delete_nav_link(link_id):
-    """Delete a navigation link."""
+def nav_link_operations(link_id):
+    """Update or delete a navigation link."""
     link = NavLink.query.filter_by(id=link_id, user_id=current_user.id).first_or_404()
-    db.session.delete(link)
+
+    if request.method == 'DELETE':
+        db.session.delete(link)
+        db.session.commit()
+        return '', 204
+
+    data = request.json
+    if 'embed' in data:
+        link.embed = data['embed']
+    if 'name' in data:
+        link.name = data['name']
+    if 'url' in data:
+        link.url = data['url']
+    if 'icon_class' in data:
+        link.icon_class = data['icon_class']
+
     db.session.commit()
-    return '', 204
+    return jsonify({
+        'id': link.id,
+        'name': link.name,
+        'url': link.url,
+        'icon_class': link.icon_class,
+        'embed': link.embed
+    })
 
 with app.app_context():
     db.create_all()
